+++ ChangeLog.md
@@ -1,6 +1,8 @@
 # Rogallo ChangeLog
 
-## Unreleased
+## v0.12.0
+
+**Released: 2026-07-20**
 
 - Added `!theme` as a command line command.
   ([#186](https://github.com/davep/rogallo/pull/186))